aggregated BingCovid and scrape_bing.py
Reference Issues/PRs
None

What does this implementation fix/change?

This change aggregates BingCovid.py and scrape_bing.py. The class BingCovid defined in BingCovid.py is now found in the aggregated root scrape_bing.py.

other comments:
 no
