--- scraping/ScrapeBingCovid/BingCovid.py
@@ -1,26 +1,0 @@
-from datetime import datetime
-from dateutil import parser
-
-
-class BingCovid:
-    def __init__(
-        self,
-        state=None,
-        country=None,
-        last_update=None,
-        lat=None,
-        lng=None,
-        confirmed=None,
-        deaths=None,
-        recovered=None,
-        posted_date=datetime.utcnow(),
-    ):
-        self.state = state
-        self.country = country
-        self.last_update = parser.parse(last_update) if last_update else last_update
-        self.lat = lat
-        self.lng = lng
-        self.confirmed = confirmed
-        self.deaths = deaths
-        self.recovered = recovered
-        self.posted_date = posted_date
