+++ scraping/ScrapeBingCovid/scrape_bing.py
@@ -25,8 +25,31 @@ API_URL = "https://bing.com/covid/data"
 # ScrapeRss helper function
 from ScrapeRss.helpers import get_seed_page
 
+from datetime import datetime
+from dateutil import parser
 # BingCovid
-from ScrapeBingCovid.BingCovid import BingCovid
+class BingCovid:
+    def __init__(
+        self,
+        state=None,
+        country=None,
+        last_update=None,
+        lat=None,
+        lng=None,
+        confirmed=None,
+        deaths=None,
+        recovered=None,
+        posted_date=datetime.utcnow(),
+    ):
+        self.state = state
+        self.country = country
+        self.last_update = parser.parse(last_update) if last_update else last_update
+        self.lat = lat
+        self.lng = lng
+        self.confirmed = confirmed
+        self.deaths = deaths
+        self.recovered = recovered
+        self.posted_date = posted_date
 
 if __name__ == "__main__":
     db_bingcovid.connect()
